Team_13_ArnavGupta_CSE1/Source-code/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
mysql-connector-python==8.4.0
PyJWT==2.8.0
bcrypt==4.1.3
redis==7.4.0
